Fix and update codeclimate and travis-ci config
The way test coverage needs to be reported to Codeclimate has changed and the version used in this repo was out of date. Also some of the config being used was either redundant or simply not the standard we use across other projects so these issues were resolved in this fix as well.