Add changelog and ability to auto generate it
To have better visibility on what is changed between versions we try to add a changelog to our open sourced gem projects.

However rather than generate it manually we rely on the [GitHub Changelog generator](https://github.com/skywinder/Github-Changelog-Generator) to do that for us.

This change adds a first iteration of the changelog and the ability to generate it going forward.