Update travis config to enable auto deployments
Where we have a gem that is shared across services, we release it to rubygems with proper versioning to allow the dependent services to control which versions they bring in.

To simplify this process we can enable Travis to automatically build and deploy the gem to rubygems.org when we push a new tag to the github repo.

The changes here update the Travis config to enable this.